+++ src/e2e/src/integration/upload-toolbar.spec.ts
@@ -1,5 +1,5 @@
 import { simulateDrop } from "../utils/drag-event";
-import { browser, logging } from "protractor";
+import { browser } from "protractor";
 import { spawn, ChildProcess } from "child_process";
 import { Dashboard } from "../support/dashboard.po";
 import { UploadToolbarPO } from "../support/ngx-fileupload-ui/upload-toolbar";
@@ -12,12 +12,6 @@ describe("Ngx Fileupload Upload Toolbar", () => {
     const ngxFileUpload: NgxFileuploadPO = new NgxFileuploadPO();
     const uploadToolbar: UploadToolbarPO = new UploadToolbarPO();
     const dashboard: Dashboard = new Dashboard();
-
-    function slowDownUpload() {
-        server.send({
-            timeout: 1000
-        });
-    }
 
     beforeAll(async () => {
         // start very simple upload server which only logs
@@ -78,7 +72,7 @@ describe("Ngx Fileupload Upload Toolbar", () => {
         /** start uploads */
         await browser.waitForAngularEnabled(false);
         await uploadToolbar.uploadAll();
-        await browser.sleep(100);
+        await browser.sleep(200);
 
         expect(
             await uploadToolbar.uploadStates.map<string>((state) => state.getText())
@@ -88,6 +82,7 @@ describe("Ngx Fileupload Upload Toolbar", () => {
     it("should contain 3 progressing, 7 pending and 1 idle in info bar", async () => {
         // add another file
         await simulateDrop(ngxFileUpload.getFileBrowser(), "./upload-file.zip");
+        await browser.sleep(200);
 
         expect(
             await uploadToolbar.uploadStates.map<string>((state) => state.getText())
@@ -113,18 +108,6 @@ describe("Ngx Fileupload Upload Toolbar", () => {
 
         /** start uploads */
         await uploadToolbar.removeAll();
-    });
-
-    afterEach(async () => {
-        // Assert that there are no errors emitted from the browser
-        let logs = await browser.manage().logs().get(logging.Type.BROWSER);
-
-        /** we expect an error in browser with state of 401 so we filter this out */
-        const error = "responded with a status of 401 (Unauthorized)";
-        logs = logs.filter((entry: logging.Entry) => !entry.message.endsWith(error));
-
-        expect(logs)
-            .not.toContain(jasmine.objectContaining({ level: logging.Level.SEVERE } as logging.Entry));
     });
 
     afterAll(() => {