+++ src/example/libs/pages/customize/src/item-template/item-template.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy, Inject } from "@angular/core";
-import { UploadStorage, UploadRequest, UploadApi } from "@r-hannuschka/ngx-fileupload";
+import { UploadStorage, UploadRequest, UploadState } from "@r-hannuschka/ngx-fileupload";
 import { Subject } from "rxjs";
 import { takeUntil } from "rxjs/operators";
 
@@ -26,7 +26,7 @@ export class ItemTemplateComponent implements OnInit, OnDestroy {
 
     public showDocs = false;
 
-    public uploadStates = UploadApi.UploadState;
+    public uploadStates = UploadState;
 
     public uploads: UploadRequest[] = [];
 
@@ -53,9 +53,5 @@ export class ItemTemplateComponent implements OnInit, OnDestroy {
         this.destroy$.next(true);
         this.destroy$.complete();
         this.destroy$ = null;
-    }
-
-    public removeUpload(requestId) {
-        this.storage.remove(requestId);
     }
 }