+++ src/lib/ngx-fileupload/data/api/index.ts
@@ -1,6 +1,0 @@
-export interface IDataNode {
-    [key: string]: any;
-}
-
-export * from "./validation";
-export * from "./upload";